// src/plantify_admin/src/declarations/plantify_backend/plantify_backend.did.d.ts
export interface Startup {
  'id' : string,
  'status' : string,
  'periodicProfitSharing' : string,
  'foundedYear' : string,
  'competitiveAdvantage' : string,
  'createdAt' : Time,
  'businessPlan' : [] | [string],
  'description' : string,
  'sector' : string,
  'useOfFunds' : string,
  'website' : string,
  'teamMembers' : Array<TeamMember>,
  'targetMarket' : string,
  'updatedAt' : Time,
  'revenueModel' : string,
  'solution' : string,
  'companyLogo' : [] | [string],
  'founderId' : string,
  'companyType' : string,
  'financialProjections' : [] | [string],
  'marketingStrategy' : string,
  'startupName' : string,
  'fundingGoal' : string,
  'legalDocuments' : [] | [string],
  'monthlyRevenue' : string,
  'operationalProcess' : string,
  'advisors' : string,
  'nftPrice' : string,
  'location' : string,
  'monthlyExpenses' : string,
  'problemStatement' : string,
  'founderBackground' : string,
}
// ...
export interface TeamMember {
  'id' : bigint,
  'linkedin' : string,
  'background' : string,
  'name' : string,
  'role' : string,
  'email' : string,
  'isFounder' : boolean,
  'photo' : [] | [string],
}
export type Time = bigint;